requirements.txt
PyGithub
python-dotenv
pandas
pyarrow
boto3
pqdm
optuna
xgboost
scikit-learn
fastapi
uvicorn[standard]
setuptools
prometheus-fastapi-instrumentator
pytest
moto[s3]==4.2.13